requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.8.6
fluent.runtime==0.3.1
fluent.syntax==0.18.1
GitPython==3.1.43
GitPython==3.1.43
pydantic==1.10.5
pymongo==4.6.3
python-dotenv==1.0.1
Requests==2.31.0
